Upgrade the shader, Fixed minor casting issues

+++ Source/Cubiquity/Classes/Cubiquity.hpp
@@ -4,7 +4,6 @@
 #define CUBIQUITY_HPP_
 
 #include "CubiquityC.h"
-#include "PhysicsEngine/BodySetup.h"
 
 #include <stdexcept>
 #include <sstream>
@@ -154,7 +153,7 @@ namespace Cubiquity
 
 		Vector<uint8_t> encodedPos() const { return{ m_vertex.encodedPosX, m_vertex.encodedPosY, m_vertex.encodedPosZ }; }
 
-		Vector<double> position() const { return{ m_vertex.encodedPosX - 0.5, m_vertex.encodedPosY - 0.5, m_vertex.encodedPosZ - 0.5 }; }
+		Vector<float> position() const { return{ (float)(m_vertex.encodedPosX - 0.5f), (float)(m_vertex.encodedPosY - 0.5f), (float)(m_vertex.encodedPosZ - 0.5f) }; }
 
 		Color color() const { return{ m_vertex.data }; };
 
@@ -175,7 +174,7 @@ namespace Cubiquity
 
 		uint16_t encodedNormal() const { return{ m_vertex.encodedNormal }; }
 
-		Vector<double> position() const { return{ (1.0 / 256.0) * m_vertex.encodedPosX, (1.0 / 256.0) * m_vertex.encodedPosY, (1.0 / 256.0) * m_vertex.encodedPosZ }; }
+		Vector<float> position() const { return{ (float)(1.0 / 256.0) * m_vertex.encodedPosX, (float)(1.0 / 256.0) * m_vertex.encodedPosY, (float)(1.0 / 256.0) * m_vertex.encodedPosZ }; }
 
 		Vector<float> normal() const
 		{
